requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
chardet==4.0.0
discord==1.7.3
discord.py==1.7.3
dnspython==2.2.1
idna==3.4
multidict==6.0.2
pycodestyle==2.9.1
pykerberos==1.2.4
python-dotenv==0.21.0
toml==0.10.2
typing_extensions==4.3.0
yarl==1.8.1
